+++ Client/src/pages/Home/RestaurantContainer.jsx
@@ -1,13 +1,8 @@
 import { Link } from "react-router-dom";
-import { deleteRestaurant } from "../../services/requests";
 
-export default function RestaurantContainer({ restaurant, onUpdate }) {
+export default function RestaurantContainer({ restaurant }) {
   console.log("Restaurant props", restaurant);
 
-  async function deleteRestaurantHandler() {
-    deleteRestaurant(restaurant.id);
-    onUpdate(restaurant.id);
-  }
   return (
     <>
       <Link to={`/restaurant/${restaurant.id}`}>
@@ -15,10 +10,6 @@ export default function RestaurantContainer({ restaurant, onUpdate }) {
           <h3>{restaurant.name}</h3>
           <p>{restaurant.description}</p>
         </div>
-      </Link>
-      <button onClick={deleteRestaurantHandler}>Delete</button>
-      <Link to={`/restaurant/${restaurant.id}/update`}>
-        <button>Update Restaurant</button>
       </Link>
     </>
   );
