+++ Client/src/pages/YourRestaurants/NewRestaurant/NewRestaurant.jsx
@@ -1,14 +1,12 @@
-import React from "react";
-import { createRestaurant, updateRestaurant } from "../../../services/requests";
-import { useNavigate, useParams } from "react-router-dom";
+import { createRestaurant } from "../../../services/requests";
+import { useNavigate } from "react-router-dom";
 
-function UpdateRestaurant() {
+function NewRestaurant() {
   const navigate = useNavigate();
-  const { id } = useParams();
+  console.log(history);
 
   async function submitRestaurantData() {
-    const restaurantData = {
-      id: id,
+    await createRestaurant({
       name: document.getElementById("restaurantForm--name").value,
       description: document.getElementById("restaurantForm--description").value,
       timings: {
@@ -17,29 +15,24 @@ function UpdateRestaurant() {
           document.getElementById("restaurantForm--close").value
         ),
       },
-    };
-    await updateRestaurant(id, restaurantData);
+    });
     navigate("/");
   }
   return (
     <div>
       <label>Name:</label>
       <input type="text" id="restaurantForm--name"></input>
-
       <label>Description:</label>
       <input type="text" id="restaurantForm--description"></input>
-
       <label>Opens At:</label>
       <input type="text" id="restaurantForm--open"></input>
-
       <label>Closes At:</label>
       <input type="text" id="restaurantForm--close"></input>
-
       <button type="submit" onClick={submitRestaurantData}>
-        Update Restaurant
+        Create Restaurant
       </button>
     </div>
   );
 }
 
-export default UpdateRestaurant;
+export default NewRestaurant;
