+++ Client/src/pages/YourRestaurants/UpdateRestaurant/UpdateRestaurant.jsx
@@ -1,13 +1,13 @@
-import React from "react";
-import { createRestaurant } from "../../../services/requests";
-import { useNavigate } from "react-router-dom";
+import { updateRestaurant } from "../../../services/requests";
+import { useNavigate, useParams } from "react-router-dom";
 
-function NewRestaurant() {
+function UpdateRestaurant() {
   const navigate = useNavigate();
-  console.log(history);
+  const { id } = useParams();
 
   async function submitRestaurantData() {
-    await createRestaurant({
+    const restaurantData = {
+      id: id,
       name: document.getElementById("restaurantForm--name").value,
       description: document.getElementById("restaurantForm--description").value,
       timings: {
@@ -16,24 +16,29 @@ function NewRestaurant() {
           document.getElementById("restaurantForm--close").value
         ),
       },
-    });
+    };
+    await updateRestaurant(id, restaurantData);
     navigate("/");
   }
   return (
     <div>
       <label>Name:</label>
       <input type="text" id="restaurantForm--name"></input>
+
       <label>Description:</label>
       <input type="text" id="restaurantForm--description"></input>
+
       <label>Opens At:</label>
       <input type="text" id="restaurantForm--open"></input>
+
       <label>Closes At:</label>
       <input type="text" id="restaurantForm--close"></input>
+
       <button type="submit" onClick={submitRestaurantData}>
-        Create Restaurant
+        Update Restaurant
       </button>
     </div>
   );
 }
 
-export default NewRestaurant;
+export default UpdateRestaurant;
